Make tests work on all platforms
The JS test runner doesn't allow for nested tests, so we had to extract them.

We can't use the `.` character in tests name or the native test runner will fail from Gradles side because it gets really wonky and thinks it's a package name or something? Don't really understand it all too well.