Refactors and optimized lexing
Addresses issue #10.

This is a fairly big change that removes a generic stateful lexer. Instead, the lexer now only has a generic concept of a matcher and a new concept of a context. The context can be practically anything - including a state to switch matchers. When a fragment is matched - the lexer rule has access to the context, and therefore it can react to and modify it.

The Jukkas lexer rules are now only initialized once. When a new JukkasLexer instance is created - only a new context is created (which is much more lightweight) and passed to the existing lexer rules.